[BEAM-3227] improvements
