+++ client/Packages/com.beamable/Runtime/InventoryViewEx.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 581cd748df03f014f90ac9cbdae5b2cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
